Destination lookup table
Added a "table" of the 12 possible destinations. It's just a dedicated block of memory with each coordinate taking 2 rows of memory. There is a subroutine for getting the next coordinate GetNextDest. When it is first called, it returns the first pair of coordinates (X0, Y0). Since the subroutine automatically increments its index, each subsequent call returns the next pair of coordinates. There is another subroutine for setting the index to a certain value. When getting coordinate n (Xn, Yn), the AC should be 2*n due to how the table is set up.

THERE IS NO BOUNDS CHECKING